+++ algo/felina/State.h
@@ -77,9 +77,9 @@ struct State {
    * Get the result of the game from perspective of first player.
    * This function can be called only when the state is terminal.
    *
-   * @return 1 if a win, 0.5 if a draw, 0 if a loss.
+   * @return 1 if a win, 0 if a draw, -1 if a loss.
    */
-  double getResult() const;
+  int getResult() const;
 
   /*
    * Get the score of the game from perspective of first player.
